Add: updating favorites with async storage functionality
